[feat] Additional logics
- Added a password encryption logic
- Added a existed id check logic

- 비밀번호 암호화 로직 추가
- 존재하는 아이디 체크 로직 추가
