+++ eatgo-common/src/main/java/kr/co/fastcampus/eatgo/domain/UserRepository.java
@@ -10,4 +10,6 @@ public interface UserRepository extends CrudRepository<User, Long> {
     List<User> findAll();
 
     Optional<User> findById(Long id);
+
+    Optional<User> findByEmail(String email);
 }
