+++ eatgo-customer-api/src/main/java/kr/co/fastcampus/eatgo/application/EmailExistedExecption.java
@@ -1,0 +1,8 @@
+package kr.co.fastcampus.eatgo.application;
+
+public class EmailExistedExecption extends RuntimeException {
+
+    EmailExistedExecption(String email){
+        super("Email is already registered: " + email);
+    }
+}
