+++ eatgo-customer-api/src/main/java/kr/co/fastcampus/eatgo/application/UserService.java
@@ -1,11 +1,13 @@
-package kr.co.fastcampus.eatgo.interfaces;
+package kr.co.fastcampus.eatgo.application;
 
 import kr.co.fastcampus.eatgo.domain.User;
 import kr.co.fastcampus.eatgo.domain.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -19,10 +21,21 @@ public class UserService {
     }
 
     public User registerUser(String email, String name, String password) {
+        // 이미 존재하는지 체크
+        Optional<User> existed = userRepository.findByEmail(email);
+        if(existed.isPresent()){
+            throw new EmailExistedExecption(email);
+        }
+
+        // 암호화
+        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+        String encodedPassword = passwordEncoder.encode(password);
+
         User user = User.builder()
                 .email(email)
                 .name(name)
-                .password(password)
+                .password(encodedPassword)
+                .level(1L)
                 .build();
         return userRepository.save(user);
     }
