--- eatgo-customer-api/src/test/java/kr/co/fastcampus/eatgo/interfaces/UserServiceTest.java
@@ -1,35 +1,0 @@
-package kr.co.fastcampus.eatgo.interfaces;
-
-import kr.co.fastcampus.eatgo.domain.UserRepository;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.verify;
-
-public class UserServiceTest {
-
-    private UserService userService;
-
-    @Mock
-    private UserRepository userRepository;
-
-    @BeforeEach
-    public void setUp(){
-        MockitoAnnotations.openMocks(this);
-        userService = new UserService(userRepository);
-    }
-
-    @Test
-    public void registerUser(){
-        String email = "tester@example.com";
-        String name = "Tester";
-        String password = "test";
-
-        userService.registerUser(email, name, password);
-
-        verify(userRepository).save(any());
-    }
-}
